chore: move infocols to /classic
