+++ src/classic/components/InfoCols/index.ts
