+++ src/classic/components/index.ts
@@ -6,6 +6,7 @@ export { default as Header } from "./Header"
 export { default as Hero } from "./Hero"
 export { default as Image } from "./Image"
 export { default as Infobar } from "./Infobar"
+export { default as InfoCols } from "./InfoCols"
 export { default as InfoCards } from "./InfoCards"
 export { default as InfoPic } from "./Infopic"
 export { default as Masthead } from "./Masthead"
