+++ src/common/index.ts
@@ -6,6 +6,7 @@ export { default as HeaderProps } from "./Header"
 export { default as HeroProps } from "./Hero"
 export { default as ImageProps } from "./Image"
 export { default as InfobarProps } from "./Infobar"
+export { default as InfoColsProps } from "./InfoCols"
 export { default as InfoCardsProps } from "./InfoCards"
 export { default as InfopicProps } from "./Infopic"
 export { default as MastheadProps } from "./Masthead"
